BackEnd/resume-analyzer/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
requests
pydantic
python-multipart